calculator
a calculator with basic mathematical functionalities using CSS, HTML, and JavaScript. Calculators require an interactive interface to perform basic functions such as addition, subtraction, division, and multiplication. It will need a  display screen to display the user’s input and give the results. The grid system in CSS is for the alignments of buttons on the calculator.